+++ src/Appwrite/Migration/Version/V18.php
@@ -82,7 +82,7 @@ class V18 extends Migration
                  */
                 $document->setAttribute('version', '1.2.0');
                 break;
-            case 'users':
+            case 'projects':
                 /**
                  * Bump version number.
                  */
@@ -92,7 +92,8 @@ class V18 extends Migration
                  * Set default passwordHistory
                  */
                 $document->setAttribute('auths', array_merge($document->getAttribute('auths', []), [
-                    'passwordHistory' => 0
+                    'passwordHistory' => 0,
+                    'passwordDictionary' => false,
                 ]));
 
                 break;